requirements.txt
fastapi
uvicorn
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
asyncpg
qdrant-client
langchain-openai
langsmith
python-multipart
pytest
google-generativeai
openai
gTTS
pydantic-settings
pydantic[email]
elevenlabs
google-cloud-texttospeech>=2.16

